fix type issues and format with black

+++ spamdb/modules/jsbot.py
@@ -7,6 +7,9 @@ def update_jsbot_colls() -> None:
     args = env.args
     db = env.db
 
+    if not args.jsbots:
+        return
+
     if args.drop:
         db.jsbot.drop()
         db.jsbot_asset.drop()
@@ -14,9 +17,10 @@ def update_jsbot_colls() -> None:
     if args.no_create:
         return
 
-    util.bulk_write(db.jsbot, _read_json("jsbot.json"))
-    util.bulk_write(db.jsbot_asset, _read_json("jsbot.asset.json"))
+    util.bulk_write(db.jsbot, _read_json('jsbot.json'))
+    util.bulk_write(db.jsbot_asset, _read_json('jsbot.asset.json'))
+
 
 def _read_json(file: str) -> list:
-    with open(f"{env.data_path}/{file}", encoding='utf-8') as f:
+    with open(f'{env.data_path}/{file}', encoding='utf-8') as f:
         return json.load(f)
